+++ backend/src/baserow/core/registries.py
@@ -554,7 +554,7 @@ class ObjectScopeType(Instance, ModelInstanceMixin):
 
     def contains(self, context: ContextObject):
         """
-        Return True if the context is one object of this context.
+        Returns True if the context is one object of this context.
 
         :param context: The context to test
         :return: True if the ObjectScopeType of the context is the same as this one.
@@ -563,8 +563,25 @@ class ObjectScopeType(Instance, ModelInstanceMixin):
         context_scope_type = object_scope_type_registry.get_by_model(context)
         return context_scope_type.type == self.type
 
+    @cached_property
+    def level(self) -> int:
+        """
+        Returns the level of this scope in the full object hierarchy. The level is the
+        number of ancestor to get to the root object.
 
-class ObjectScopeTypeRegistry(Registry[ObjectScopeType], ModelRegistryMixin):
+        :return: The level of the scope.
+        """
+
+        parent = self.get_parent_scope()
+        if parent is None:
+            return 0
+        else:
+            return parent.level + 1
+
+
+class ObjectScopeTypeRegistry(
+    Registry[ObjectScopeType], ModelRegistryMixin[Any, ObjectScopeType]
+):
     """
     This registry contains all `ObjectScopeType`. It also proposes a set of methods
     useful to go through the full object/scope hierarchy.
@@ -644,7 +661,7 @@ class ObjectScopeTypeRegistry(Registry[ObjectScopeType], ModelRegistryMixin):
         if child_scope_type is None:
             return False
 
-        if parent_scope_type == child_scope_type:
+        if parent_scope_type.type == child_scope_type.type:
             return True
         else:
             return self.scope_type_includes_scope_type(
@@ -686,7 +703,7 @@ class SubjectType(Instance, ModelInstanceMixin):
         )
 
 
-class SubjectTypeRegistry(Registry[SubjectType], ModelRegistryMixin):
+class SubjectTypeRegistry(Registry[SubjectType], ModelRegistryMixin[Any, SubjectType]):
     """
     This registry holds all the different subject types used across Baserow.
     """