+++ enterprise/backend/tests/baserow_enterprise_tests/role/test_role_handler.py
@@ -617,3 +617,112 @@ def test_assign_role_batch_performance(data_fixture):
 
     print("----------Database Level Deletions------------")
     print(profiler.output_text(unicode=True, color=True))
+
+
+@pytest.mark.django_db(transaction=True)
+def test_get_sorted_subject_scope_roles(data_fixture, enterprise_data_fixture):
+    user = data_fixture.create_user()
+    user2 = data_fixture.create_user()
+    group = data_fixture.create_group(user=user, members=[user2])
+    database1 = data_fixture.create_database_application(user=user, group=group)
+    table11 = data_fixture.create_database_table(user=user, database=database1)
+    table12 = data_fixture.create_database_table(user=user, database=database1)
+    database2 = data_fixture.create_database_application(user=user, group=group)
+    table21 = data_fixture.create_database_table(user=user, database=database2)
+    table22 = data_fixture.create_database_table(user=user, database=database2)
+
+    team1 = enterprise_data_fixture.create_team(group=group, members=[user2])
+    team2 = enterprise_data_fixture.create_team(group=group, members=[user2])
+    team3 = enterprise_data_fixture.create_team(group=group, members=[user2])
+
+    admin_role = Role.objects.get(uid="ADMIN")
+    editor_role = Role.objects.get(uid="EDITOR")
+    builder_role = Role.objects.get(uid="BUILDER")
+    viewer_role = Role.objects.get(uid="VIEWER")
+    no_role_role = Role.objects.get(uid="NO_ACCESS")
+    low_priority_role = Role.objects.get(uid="NO_ROLE_LOW_PRIORITY")
+
+    RoleAssignmentHandler().assign_role(user2, group, role=editor_role, scope=group)
+
+    assert RoleAssignmentHandler().get_roles_per_scope(group, user2) == [
+        (group, [editor_role])
+    ]
+
+    RoleAssignmentHandler().assign_role(user2, group, role=viewer_role, scope=table11)
+
+    assert RoleAssignmentHandler().get_roles_per_scope(group, user2) == [
+        (group, [editor_role]),
+        (table11, [viewer_role]),
+    ]
+
+    RoleAssignmentHandler().assign_role(
+        user2, group, role=builder_role, scope=database1
+    )
+
+    assert RoleAssignmentHandler().get_roles_per_scope(group, user2) == [
+        (group, [editor_role]),
+        (database1, [builder_role]),
+        (table11, [viewer_role]),
+    ]
+
+    RoleAssignmentHandler().assign_role(user2, group, role=editor_role, scope=database2)
+    RoleAssignmentHandler().assign_role(user2, group, role=builder_role, scope=table22)
+
+    assert RoleAssignmentHandler().get_roles_per_scope(group, user2) == [
+        (group, [editor_role]),
+        (database1, [builder_role]),
+        (database2, [editor_role]),
+        (table11, [viewer_role]),
+        (table22, [builder_role]),
+    ]
+
+    RoleAssignmentHandler().assign_role(team1, group, role=editor_role, scope=group)
+    RoleAssignmentHandler().assign_role(team2, group, role=builder_role, scope=group)
+    RoleAssignmentHandler().assign_role(team3, group, role=no_role_role, scope=group)
+
+    RoleAssignmentHandler().assign_role(team1, group, role=editor_role, scope=database1)
+
+    assert RoleAssignmentHandler().get_roles_per_scope(group, user2) == [
+        (group, [editor_role]),
+        (database1, [builder_role]),
+        (database2, [editor_role]),
+        (table11, [viewer_role]),
+        (table22, [builder_role]),
+    ]
+
+    RoleAssignmentHandler().assign_role(
+        team2, group, role=builder_role, scope=database1
+    )
+
+    assert RoleAssignmentHandler().get_roles_per_scope(group, user2) == [
+        (group, [editor_role]),
+        (database1, [builder_role]),
+        (database2, [editor_role]),
+        (table11, [viewer_role]),
+        (table22, [builder_role]),
+    ]
+
+    RoleAssignmentHandler().assign_role(
+        user2, group, role=low_priority_role, scope=group
+    )
+
+    assert RoleAssignmentHandler().get_roles_per_scope(group, user2) == [
+        (group, [editor_role, builder_role, no_role_role]),
+        (database1, [builder_role]),
+        (database2, [editor_role]),
+        (table11, [viewer_role]),
+        (table22, [builder_role]),
+    ]
+
+    RoleAssignmentHandler().assign_role(team1, group, role=editor_role, scope=table12)
+    RoleAssignmentHandler().assign_role(team2, group, role=builder_role, scope=table12)
+    RoleAssignmentHandler().assign_role(team3, group, role=no_role_role, scope=table12)
+
+    assert RoleAssignmentHandler().get_roles_per_scope(group, user2) == [
+        (group, [editor_role, builder_role, no_role_role]),
+        (database1, [builder_role]),
+        (database2, [editor_role]),
+        (table11, [viewer_role]),
+        (table12, [editor_role, builder_role, no_role_role]),
+        (table22, [builder_role]),
+    ]