+++ enterprise/web-frontend/modules/baserow_enterprise/components/MembersRoleField.vue
@@ -36,6 +36,7 @@ import { clone } from '@baserow/modules/core/utils/object'
 import { notifyIf } from '@baserow/modules/core/utils/error'
 import RoleAssignmentsService from '@baserow_enterprise/services/roleAssignments'
 import EditRoleContext from '@baserow/modules/core/components/settings/members/EditRoleContext'
+import { filterRoles } from '@baserow_enterprise/utils/roles'
 
 export default {
   name: 'MembersRoleField',
@@ -57,7 +58,12 @@ export default {
       )
     },
     roles() {
-      return this.group ? this.group._.roles : []
+      return this.group
+        ? filterRoles(this.group._.roles, {
+            scopeType: 'group',
+            subjectType: 'auth.User',
+          })
+        : []
     },
     ...mapGetters({ userId: 'auth/getUserId' }),
   },