+++ enterprise/web-frontend/modules/baserow_enterprise/enums/roles.js
@@ -1,8 +1,0 @@
-// TODO delete once roles are fetched from endpoint
-export const roles = [
-  { uid: 'ADMIN' },
-  { uid: 'BUILDER' },
-  { uid: 'EDITOR' },
-  { uid: 'COMMENTER' },
-  { uid: 'VIEWER' },
-]